requirements.txt
requests
anthropic
python-dotenv
feedparser
